feature: 사타리 게임에 필요한 정보를 담는다.
